requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14,<4.0
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0
Pillow>=10.0,<13.0
openpyxl>=3.1,<4.0
python-barcode>=0.15,<1.0
reportlab>=4.0,<5.0
